+++ magic_bridge/ic/DIP20/.gitignore
@@ -1,6 +1,6 @@
 .DS_Store
 .dfx/
 build/
-target
+target/
 .config
 .cache
