+++ magic_bridge/ic/DIP20/.gitmodules
@@ -1,0 +1,3 @@
+[submodule "rust/cap"]
+	path = rust/cap
+	url = https://github.com/Psychedelic/cap.git
