+++ magic_bridge/ic/DIP20/rust/Cargo.toml
@@ -1,19 +1,4 @@
-[package]
-name = "token"
-version = "0.1.0"
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-ic-cdk-macros = "0.3"
-candid = "0.7.4"
-serde = "1.0"
-ic-kit = "0.4.3"
-ic-cdk = "0.3.1"
-assert-panic = "1.0.1"
-cap-std = { git = "https://github.com/Psychedelic/cap", branch = "main", package="cap-standards", features = ["alpha-dip20", "cap-sdk", "sdk-impls"] }
-cap-sdk = { git = "https://github.com/Psychedelic/cap.git", branch = "main" }
-
-[target.'cfg(not(target_family = "wasm"))'.dependencies]
-async-std = { version="1.10.0", features = ["attributes"] }
+[workspace]
+members = [
+    "token",
+]
