+++ magic_bridge/ic/DIP20/rust/token/token.did
@@ -23,9 +23,9 @@ type TxError = variant {
   Unauthorized;
   LedgerTrap;
   ErrorTo;
+  Other : text;
   BlockUsed;
   AmountTooSmall;
-  Other : text;
 };
 service : (
   text,
@@ -63,5 +63,3 @@ service : (
   transfer : (principal, nat) -> (Result);
   transferFrom : (principal, principal, nat) -> (Result);
 }
-
-dfx canister --network ic call 6trho-eyaaa-aaaab-qafnq-cai totalSupply '()'
